requirements.txt
M2Crypto==0.33.0
redis==3.2.1
mysqlclient==1.4.2.post1
celery==4.2.1
Django==1.11.15
django-bootstrap3==10.0.1
django-filter==1.1.0
django-tables2==1.21.2
lxml==4.2.4
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
requests-toolbelt==0.8.0
pdfkit==0.6.1
xmltodict==0.11.0
openpyxl==2.5.6
python-dateutil==2.7.3
qrcode==6.0
django-celery-results==1.0.1
django-celery-beat==1.4.0
uWSGI==2.0.18
boto3==1.13.25
pycurl==7.43.0.5